fix: move `ProofExpr` to `mod.rs` due to `clippy::module-inception`
